tests/requirements.txt
pytest
pytest-mock
gitpython
PyGithub